+++ src/xcode/ENA/ENA/Source/Scenes/QRScanner/QRScannerCoordinator.swift
@@ -335,7 +335,7 @@ class QRScannerCoordinator {
 		_ initializationData: TicketValidationInitializationData
 	) {
 		showActivityIndicator()
-		let ticketValidation = MockTicketValidation(with: initializationData)
+		let ticketValidation = MockTicketValidation(with: initializationData, restServiceProvider: .fake())
 		ticketValidation.delay = 1
 
 		ticketValidation.initialize { [weak self] result in
