+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/TicketValidation.swift
@@ -9,9 +9,11 @@ final class TicketValidation: TicketValidating {
 	// MARK: - Protocol TicketValidating
 
 	init(
-		with initializationData: TicketValidationInitializationData
+		with initializationData: TicketValidationInitializationData,
+		restServiceProvider: RestServiceProviding
 	) {
 		self.initializationData = initializationData
+		self.restServiceProvider = restServiceProvider
 	}
 
 	var initializationData: TicketValidationInitializationData
@@ -42,4 +44,31 @@ final class TicketValidation: TicketValidating {
 	func cancel() {
 
 	}
+	
+	func validateIdentityDocumentOfValidationDecorator(
+		urlString: String,
+		completion:
+		@escaping (Result<ServiceIdentityDocumentValidationDecorator, ServiceIdentityValidationDecoratorError>) -> Void
+	) {
+		guard let url = URL(string: urlString) else {
+			Log.error("URL cant be constructed from input string", log: .ticketValidationDecorator)
+			return
+		}
+		let resource = ServiceIdentityDocumentValidationDecoratorResource(url: url)
+		restServiceProvider.load(resource) { result in
+			switch result {
+			case .success(let model):
+				TVDecoratorIdentityDocumentProcessor().validateIdentityDocument(serviceIdentityDocument: model) { result in
+					completion(result)
+				}
+			case .failure(let error):
+				completion(.failure(.REST_SERVICE_ERROR(error)))
+				Log.error(error.localizedDescription, log: .ticketValidationDecorator)
+			}
+		}
+	}
+	
+	// MARK: - Private
+	
+	private let restServiceProvider: RestServiceProviding
 }
